Add support for encoding X500DistinguishedName values on Unix.
The Unix implementation matches the Windows behavior on the Use*
flags for choosing the RDN separator, but always uses the UTF-8
encoding instead of the UCS-2/BMP encoding (Windows will use UTF-8 as
an opt-in with a flag).

So new X500DistinguishedName(input, flags).Decode(false) will always
produce the same output string from both Windows and Unix, but the RawData
will differ.