Support for loading arbitrary curves into OpenSSL.
This could be used to enable support of secp224r1 on CentOS, but it would
probably be fragmented (not work from NamedCurve certificate public keys).

So, instead, this is more likely to serve as support for a future arbitrary
curve feature.

Some aspects, such as known curve extraction, may be more useful in the short-term.

And the EcGroup functions should be moved to a pal_ecgroup.cpp, and all the functions
documented in header files.