Block zero-byte calls to encrypt for Unix SslStream
SSL_write is described as having an undefined behavior.
The observed behavior is that it returns 0 (which is ambiguous for "wrote 0/0
bytes" or "writing failed"), and that we then try to read the framed message and
the message buffer is empty.

Trying to detect this case once inside the PAL seemed to violate assumptions of the common code, so instead a PAL-capability check is done in
SslStreamInternal.StartWriting.  Windows will continue to write the framed
empty message, non-Windows will do nothing.