Assert that _maxDataSize is positive to prevent an infinite loop.
When setting up a PAL implementation it's really easy to just let all metadata
values be their defaults.  A MaximumMessageSize value of zero results in
data not ever getting copied (because it keeps trying to copy it into the empty array),
so add an assert in the common layer to mark that these values are a starting
point in the development of a new PAL implementation.