Make client certs and cert callbacks throw PNSE on macOS.
The darwinssl backend for libcurl does not support passing in cert pointers,
making client cert usage difficult.

The darwinssl backend also doesn't allow for any validation to occur between
the time that the handshake presents a server cert and the time that the request
headers (including cookies and the query) are sent.  Piggy-backing on the write
callback would allow validation to occur before sending a post body, but that
callback does not fire on GET or HEAD requests.