Implement X509Certificate.GetContentType in fully managed on macOS
SecItemImport can load data and identifies the load format, but it gives
inconsistent results.

If PKCS#7 is tried first, TestData.MsCertificate identifies as PKCS#7.
If X509 is tried first, some PKCS#7 blobs identify as X509.

When X509 is tried first, some PEM-encoded PKCS#7 whose DER identified as
PKCS#7 identify as X509.

Do enough scanning to make a strong plausible case, but (hopefully) not so
much as to reject legitimate inputs that the underlying system would load.