Generate RSA keys via a temporary keychain.
When no keychain is given, SecKeyGeneratePair will generate the keys into the
default/login keychain.  Since the keys were not expected to have been in the
keychain, they permanently leak out and the keychain grows over time.

Instead, have the managed code provide a temporary keychain into which the keys
will be created. Then, export them and re-import them to the NULL keychain. The
temporary keychain can then be deleted, and the hard problem of delayed cleanup is
avoided.

Files on disk will still leak if the process terminates during the key generation
phase, but that is a much smaller window than other 'temporary file on disk' solutions.